+++ src/db/guitar/main.js
@@ -1,7 +1,7 @@
 export default {
   strings: 6,
   fretsOnChord: 4,
-  name: 'Guitar',
+  name: 'guitar',
   keys: [ 'C', 'C#', 'D', 'Eb', 'E', 'F', 'F#', 'G', 'Ab', 'A', 'Bb', 'B' ],
   tunnings: {
     standard: [ 'E2', 'A2', 'D3', 'G3', 'B3', 'E4' ]