+++ src/db/ukelele/main.js
@@ -1,7 +1,7 @@
 export default {
   strings: 4,
   fretsOnChord: 4,
-  name: 'Ukelele',
+  name: 'ukelele',
   type: 'soprano',
   keys: [ 'C', 'Db', 'D', 'Eb', 'E', 'F', 'Gb', 'G', 'Ab', 'A', 'Bb', 'B' ],
   tunnings: {